requirement.txt
django==5.0.6
djangorestframework==3.15.2
drf-yasg
setuptools
